Activate function 0 bus mastering
The reason why the msleep(20) hack was required was because the device
only cares whether function 0 is a bus master or not; other function's
bus mastering status is completely ignored.

Function 0 is the NVMe drive. The NVMe drive defers suspend to a task
and as a result the NVMe drive's bus mastering isn't enabled yet when
our driver tries to start the resume process. As a note, even if the
NVMe drive didn't defer the resume, this could still be troublesome as
this would be a race (linux starts resume on separate devices
asynchronically).

This fix enables the function 0's bus master status by enabling the
device (prevents sudden bus mastering disable) and explicitly calling
pci_set_master as needed. However, enabling the device like this breaks
the NVMe driver as-is, so a separate kernel patch for the NVMe driver is
required. As the patch isn't upstreamed yet, the device enablement is
currently disabled (but can be enabled by removing the
WITHOUT_NVME_PATCH option from Makefile).